+++ test/unit/widgets/world/RenderSettingsWidgetTest.cpp
@@ -72,6 +72,7 @@ namespace RenderSettingsWidgetTest {
     EXPECT_EQ(256, widget.shadowMapSize());
     EXPECT_EQ(1, widget.shadowCascadeCount());
     EXPECT_DOUBLE_EQ(0.001, widget.shadowBias());
+    EXPECT_DOUBLE_EQ(0.0, widget.shadowSlopeBias());
     EXPECT_EQ(0, widget.shadowFilterRadius());
     EXPECT_EQ(QString("PCF"), widget.shadowFilterMode());
   }
@@ -121,12 +122,14 @@ namespace RenderSettingsWidgetTest {
     auto shadowMapSize = widget.findChild<QSpinBox*>("rasterShadowMapSize");
     auto cascadeCount = widget.findChild<QSpinBox*>("rasterShadowCascadeCount");
     auto shadowBias = widget.findChild<QDoubleSpinBox*>("rasterShadowBias");
+    auto shadowSlopeBias = widget.findChild<QDoubleSpinBox*>("rasterShadowSlopeBias");
     auto filterRadius = widget.findChild<QSpinBox*>("rasterShadowFilterRadius");
     auto filterMode = widget.findChild<QComboBox*>("rasterShadowFilterMode");
     ASSERT_NE(nullptr, shadowMaps);
     ASSERT_NE(nullptr, shadowMapSize);
     ASSERT_NE(nullptr, cascadeCount);
     ASSERT_NE(nullptr, shadowBias);
+    ASSERT_NE(nullptr, shadowSlopeBias);
     ASSERT_NE(nullptr, filterRadius);
     ASSERT_NE(nullptr, filterMode);
 
@@ -134,6 +137,7 @@ namespace RenderSettingsWidgetTest {
     shadowMapSize->setValue(1024);
     cascadeCount->setValue(3);
     shadowBias->setValue(0.25);
+    shadowSlopeBias->setValue(0.03);
     filterRadius->setValue(3);
     filterMode->setCurrentText("PCSS");
 
@@ -141,6 +145,7 @@ namespace RenderSettingsWidgetTest {
     EXPECT_EQ(1024, widget.shadowMapSize());
     EXPECT_EQ(3, widget.shadowCascadeCount());
     EXPECT_DOUBLE_EQ(0.25, widget.shadowBias());
+    EXPECT_DOUBLE_EQ(0.03, widget.shadowSlopeBias());
     EXPECT_EQ(3, widget.shadowFilterRadius());
     EXPECT_EQ(QString("PCSS"), widget.shadowFilterMode());
   }
@@ -189,6 +194,7 @@ namespace RenderSettingsWidgetTest {
     auto shadowMapSize = widget.findChild<QSpinBox*>("rasterShadowMapSize");
     auto cascadeCount = widget.findChild<QSpinBox*>("rasterShadowCascadeCount");
     auto shadowBias = widget.findChild<QDoubleSpinBox*>("rasterShadowBias");
+    auto shadowSlopeBias = widget.findChild<QDoubleSpinBox*>("rasterShadowSlopeBias");
     auto filterRadius = widget.findChild<QSpinBox*>("rasterShadowFilterRadius");
     auto filterMode = widget.findChild<QComboBox*>("rasterShadowFilterMode");
     ASSERT_NE(nullptr, engineType);
@@ -196,6 +202,7 @@ namespace RenderSettingsWidgetTest {
     ASSERT_NE(nullptr, shadowMapSize);
     ASSERT_NE(nullptr, cascadeCount);
     ASSERT_NE(nullptr, shadowBias);
+    ASSERT_NE(nullptr, shadowSlopeBias);
     ASSERT_NE(nullptr, filterRadius);
     ASSERT_NE(nullptr, filterMode);
 
@@ -203,6 +210,7 @@ namespace RenderSettingsWidgetTest {
     EXPECT_TRUE(shadowMapSize->isHidden());
     EXPECT_TRUE(cascadeCount->isHidden());
     EXPECT_TRUE(shadowBias->isHidden());
+    EXPECT_TRUE(shadowSlopeBias->isHidden());
     EXPECT_TRUE(filterRadius->isHidden());
     EXPECT_TRUE(filterMode->isHidden());
 
@@ -210,6 +218,7 @@ namespace RenderSettingsWidgetTest {
     EXPECT_FALSE(shadowMapSize->isHidden());
     EXPECT_FALSE(cascadeCount->isHidden());
     EXPECT_FALSE(shadowBias->isHidden());
+    EXPECT_FALSE(shadowSlopeBias->isHidden());
     EXPECT_FALSE(filterRadius->isHidden());
     EXPECT_FALSE(filterMode->isHidden());
   }